### 2024-11-24 aa504ae
Testfunction for total executions
