### 2024-12-01 #7466 @arfon
@john-hawkins – apologies for the slow response here. It looks like JOSS is erroneously picking up your `paper.tex` file here rather than the `paper.md`. A couple of options:

1. Prepare a new branch with only the files needed for the paper compilation (or with the `paper.tex` file removed)
2. Rename `paper.tex` to something else. e.g., `otherpaper.tex`
