+++ docs/paper/refs.bib
@@ -244,4 +244,13 @@ keywords = {CRISP-DM, Literature Review, Data Mining, Process Methodology, Deplo
 abstract = {CRISP-DM is the de-facto standard and an industry-independent process model for applying data mining projects. Twenty years after its release in 2000, we would like to provide a systematic literature review of recent studies published in IEEE, ScienceDirect and ACM about data mining use cases applying CRISP-DM. We give an overview of the research focus, current methodologies, best practices and possible gaps in conducting the six phases of CRISP-DM. The main findings are that CRISP-DM is still a de-factor standard in data mining, but there are challenges since the most studies do not foresee a deployment phase. The contribution of our paper is to identify best practices and process phases in which data mining analysts can be better supported. Further contribution is a template for structuring and releasing CRISP-DM studies.}
 }
 
-
+@inproceedings{Bugacov2017,
+   author = {Bugacov, Alejandro and Czajkowski, Karl and Kesselman, Carl and Kumar, Anoop and Schuler, Robert and Tangmunarunkit, Hongsuda},
+   year = {2017},
+   month = {10},
+   pages = {79-88},
+   title = {Experiences with DERIVA: An Asset Management Platform for Accelerating eScience},
+   volume = {2017},
+   journal = {Proceedings of the IEEE 13th International Conference on eScience (e-science)},
+   doi = {10.1109/eScience.2017.20}
+}
