### 2024-12-19 #7466 @john-hawkins
@arfon -- Sorry for my slow response. I have followed the second suggestion and renamed the latex file for an extended piece of work on applications in escience.
