### 2024-12-29 #7466 @arfon
@Wytamma @claresloggett – :wave: would you be willing to review this submission for JOSS?

Based on your experience and contributions at PyCon AU, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
