### 2025-01-16 #7672 @editorialbot
[REVIEW]: Projit: An Open Source tool for Decoupled Data Science
**Submitting author:** @john-hawkins (<a href="http://orcid.org/0000-0001-6507-3671">John Hawkins</a>)
**Repository:** https://github.com/john-hawkins/projit
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
