### 2025-02-17 john-hawkins/projit#5 @claresloggett
Python version dependency?
Hi authors - I can see the dependencies listed in requirements.txt, but I'm wondering if there is any guidance on minimum required, or maximum, Python versions?
