### 2025-02-17 #7672 @claresloggett
I added a small question about dependencies (specifically Python, which I couldn't spot) in the repo issues: https://github.com/john-hawkins/projit/issues/5
