### 2025-02-17 john-hawkins/projit#6 @claresloggett
JOSS review: API docs
Re review https://github.com/openjournals/joss-reviews/issues/7672

Just looking at the documentation: I can see the docs at https://projit.readthedocs.io/en/latest/ and I can see that under Usage, https://projit.readthedocs.io/en/latest/usage.html , there are some examples given both for CLI and Python API usage.

I think there ought to also be some systematic API documentation, i.e. a reference list of available functions and their available parameters. Does this exist anywhere?

I'm a little confused by some of the CLI help text but I will start a separate thread with questions about that to avoid confusing API docs and CLI docs, since the answers might be quite different.
