### 2025-02-17 #7672 @claresloggett
Started a repo thread about API docs at https://github.com/john-hawkins/projit/issues/6
