### 2025-02-19 john-hawkins/projit#7 @Wytamma
Better CLI errors
https://github.com/openjournals/joss-reviews/issues/7672

Hi @john-hawkins! Nice work on `projit`. I'm adding some issues relating to publication in JOSS.

I missed the `train` sub command when I tried to add a dataset to my project and the error wasn't helpful.

```bash
> projit add dataset tests/data/test.csv
usage: projit add [-h] asset name path
projit add: error: the following arguments are required: path
```
