### 2025-02-19 john-hawkins/projit#8 @Wytamma
example projit projects
https://github.com/openjournals/joss-reviews/issues/7672

Please include some example projects i.e. allow a user to clone and explore a complete project.
