### 2025-02-19 4ee160a
remove double `https://doi.org` from doi
