+++ docs/paper/refs.bib
@@ -158,7 +158,7 @@
  address = {Jeju, Korea},
  editor = {},
  ISBN = {},
- doi = {https://doi.org/10.1145/3676581.3676582}
+ doi = {10.1145/3676581.3676582}
 }
 
 @inproceedings{Hawkins:2023,
