### 2025-04-01 #7672 @john-hawkins
Thanks @arfon -- I am seeing these messages now.

Apologies, I am still new to this as a review process.
I need to change my config to get automated updates.
