### 2025-04-01 john-hawkins/projit#7 @john-hawkins
Thanks @Wytamma
I see, the CLI interpreter is assuming that the missing argument is the final one: path

But in your case it was one of the intermediate argument: the unique name for the dataset
I will modify the feedback to explicitly ask users to check for all required arguments.
