### 2025-04-01 john-hawkins/projit#6 @john-hawkins
Thank you @claresloggett

I have used the sphinx automated documentation to create docstring based documentation for all functions in the package. You can see this in the Index link from the main page.

https://projit.readthedocs.io/en/latest/genindex.html

Granted, this is bit of an unituitive way to find it.
I could explicitly link to the core functions as well as this overall index from the usage page.

Do you think that would be sufficient?

John
