### 2025-04-28 4f6ac08
adding minimal python dependency
