### 2025-04-28 john-hawkins/projit#5 @john-hawkins
Thank you @claresloggett

I have been testing using python 3.8 forward.

I have just added this requirement to the requirements.txt file
