+++ requirements.txt
@@ -1,4 +1,3 @@
-python>=3.8
 numpy>=1.16.4
 pandas>=0.25.3
 datatest>=0.11.1
