### 2025-04-29 john-hawkins/projit#6 @claresloggett
Hi @john-hawkins , I'm sorry for the very long delay on this, I've been traveling but should have checked in to let you know.

I think as an API reference, a directory of functions like that is sufficient! A couple of things that I would suggest are needed to complete the reference docs:

* It should be easy to find this index from the relevant parts of the documentation. Apologies if I've missed something here! I approached it as naively as possible, i.e. I went to https://projit.readthedocs.io/en/latest/ and then clicked on "API" in the top-level menu on the left, and then clicked around a bit. I know you said you linked to it from the main page but I am not sure which page you mean, and am failing to spot it. Either way I think it would be helpful to also be linked to from the API section of the docs.
* Some of the functions seem to be listed without any documentation - maybe they are just missing docstrings? It is also possible that some of these functions are actually unused and could be removed, I haven't checked which is the case. A couple of examples are `get_path_to_dataset()` and `get_param()`.

I am still finding it quite hard to figure out how to do some things in projit. I don't think the problem is the API reference however. For instance, I went looking for how to list datasets using the API; it is not I think mentioned in the usage guide, so I guessed it might be a function like `list_datasets()` and checked the API reference, but there is no likely-looking function. It turns out that I can list datasets just by examining the `datasets` property on the Projit object, but since this isn't a function and doesn't have a docstring it gets no automatically-generated docs. I would say in that particular case the problem is not really the API reference (although a wrapper function might be a good idea just to be consistent) but the usage guide, which doesn't quite cover an entire end-to-end use-case. I will open another issue and suggest this.
