### 2025-04-29 john-hawkins/projit#6 @claresloggett
@john-hawkins I've just realised that there is in fact a link to the API docs from the API section - it just doesn't look like a link, and the subheading formatting isn't working. So really all I have to suggest for my first point in my last comment is a formatting fix!
