### 2025-04-29 john-hawkins/projit#5 @claresloggett
Hi @john-hawkins , it's probably sufficient just to mention the required Python version in the installation instructions. That's all I had in mind with my question!

It would also be possible to upgrade to a more complete package manager and include (for example) a pyproject.toml, if you wanted to.
