### 2025-04-29 6371bac
Fixing DOI issues flagged in JOSS review
