+++ docs/paper/refs.bib
@@ -28,7 +28,8 @@
   year = {2018},
   month = {04},
   pages = {},
-  title = {DeepDIVA: A Highly-Functional Python Framework for Reproducible Experiments}
+  title = {DeepDIVA: A Highly-Functional Python Framework for Reproducible Experiments},
+  doi = {10.1109/ICFHR-2018.2018.00080}
 }
 
 @article{Miksa:2017,
@@ -162,12 +163,13 @@
 }
 
 @inproceedings{Hawkins:2023,
-   author = {John Hawkins},
-   year = {2023},
-   title = {What's in a Domain? Anaylsis of URL Features},
-   booktitle = {4th International Conference on Data Science and Cloud Computing (DSCC 2023)},
-   month = {March},
-   pages = {85-93}
+ author = {John Hawkins},
+ year = {2023},
+ title = {What's in a Domain? Anaylsis of URL Features},
+ booktitle = {4th International Conference on Data Science and Cloud Computing (DSCC 2023)},
+ month = {March},
+ pages = {85-93},
+ doi = {10.5121/csit.2023.131409}
 }
 
 @article{Reznik2022,
@@ -230,18 +232,17 @@
 }
 
 @article{SCHROER2021526,
-title = {A Systematic Literature Review on Applying CRISP-DM Process Model},
-journal = {Procedia Computer Science},
-volume = {181},
-pages = {526-534},
-year = {2021},
-note = {CENTERIS 2020 - International Conference on ENTERprise Information Systems / ProjMAN 2020 - International Conference on Project MANagement / HCist 2020 - International Conference on Health and Social Care Information Systems and Technologies 2020, CENTERIS/ProjMAN/HCist 2020},
-issn = {1877-0509},
-doi = {https://doi.org/10.1016/j.procs.2021.01.199},
-url = {https://www.sciencedirect.com/science/article/pii/S1877050921002416},
-author = {Christoph Schröer and Felix Kruse and Jorge Marx Gómez},
-keywords = {CRISP-DM, Literature Review, Data Mining, Process Methodology, Deployment},
-abstract = {CRISP-DM is the de-facto standard and an industry-independent process model for applying data mining projects. Twenty years after its release in 2000, we would like to provide a systematic literature review of recent studies published in IEEE, ScienceDirect and ACM about data mining use cases applying CRISP-DM. We give an overview of the research focus, current methodologies, best practices and possible gaps in conducting the six phases of CRISP-DM. The main findings are that CRISP-DM is still a de-factor standard in data mining, but there are challenges since the most studies do not foresee a deployment phase. The contribution of our paper is to identify best practices and process phases in which data mining analysts can be better supported. Further contribution is a template for structuring and releasing CRISP-DM studies.}
+ title = {A Systematic Literature Review on Applying CRISP-DM Process Model},
+ journal = {Procedia Computer Science},
+ volume = {181},
+ pages = {526-534},
+ year = {2021},
+ note = {CENTERIS 2020 - International Conference on ENTERprise Information Systems / ProjMAN 2020 - International Conference on Project MANagement / HCist 2020 - International Conference on Health and Social Care Information Systems and Technologies 2020, CENTERIS/ProjMAN/HCist 2020},
+ issn = {1877-0509},
+ doi = {10.1016/j.procs.2021.01.199},
+ url = {https://www.sciencedirect.com/science/article/pii/S1877050921002416},
+ author = {Christoph Schröer and Felix Kruse and Jorge Marx Gómez},
+ keywords = {CRISP-DM, Literature Review, Data Mining, Process Methodology, Deployment},
 }
 
 @inproceedings{Bugacov2017,
